liquid-api-develop/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
Markdown==3.3.6
psycopg2==2.9.3
PyJWT==2.4.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
sqlparse==0.4.2
stripe==3.2.0
tzdata==2022.1
urllib3==1.26.9
